Porting vgg
